Moving to H2 SQL scripts instead of StoredProcs for testing.

+++ src/main/java/com/fletch22/orb/command/transaction/RollbackTransactionService.java
@@ -40,7 +40,7 @@ public class RollbackTransactionService {
 	
 	public void rollbackToSpecificTransaction(BigDecimal tranId) {
 		
-		List<UndoActionBundle> undoActionBundleList = logActionService.getUndoActions(tranId.longValue());
+		List<UndoActionBundle> undoActionBundleList = logActionService.getUndoActions(tranId);
 		
 		OperationResult operationResult = null;
 		for (UndoActionBundle undoActionBundle : undoActionBundleList) {
